Finished view pages for each spec type.

+++ PcSales/Models/StorageSpec.cs
@@ -7,7 +7,7 @@ namespace PcSales.Models
         public string Manufacturer { get; set; }
         public string PartName { get; set; }
         public string Interface { get; set; }
-        public double Capacity { get; set; }
+        public int Capacity { get; set; }
         public string Type { get; set; }
     }
 }
